update time series primitives
