+++ featuretools/primitives/standard/transform_primitives/time_series_primitives/__init__.py
@@ -1,0 +1,18 @@
+from featuretools.primitives.standard.transform_primitives.time_series_primitives.rolling_count import (
+    RollingCount,
+)
+from featuretools.primitives.standard.transform_primitives.time_series_primitives.rolling_max import (
+    RollingMax,
+)
+from featuretools.primitives.standard.transform_primitives.time_series_primitives.rolling_mean import (
+    RollingMean,
+)
+from featuretools.primitives.standard.transform_primitives.time_series_primitives.rolling_min import (
+    RollingMin,
+)
+from featuretools.primitives.standard.transform_primitives.time_series_primitives.rolling_std import (
+    RollingSTD,
+)
+from featuretools.primitives.standard.transform_primitives.time_series_primitives.rolling_trend import (
+    RollingTrend,
+)
